Execute some code after the DOM is ready
Main.js contained some code that was executed immediately which did not
work since the page has not completely been loaded. This included getting the audio element,
setting its src and calling the tekstiotsikko method.
This code is now executed when the page is ready. This is done with JQuery.