Set first play button click handler in javascript not html
The play / pause button's first click event handler is now set in Javascript.
Previously it was set in the html with the onclick attribute. This caused play and
pause event handlers to be executed twice when button was
pressed. Pause can still be called multiple times
and this needs to be fixed.